pyproject.toml
[project]
name = "practical-ai-in-biology-book"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jinja2>=3.1.6",
    "markdown-it-py>=3.0.0",
    "mdit-py-plugins>=0.4.2",
    "playwright>=1.54.0",
]
